% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{addFootprintDatabyPeriod}
\alias{addFootprintDatabyPeriod}
\title{Add Footprint level data by period}
\usage{
addFootprintDatabyPeriod(input_data, metric_cols = media_names_vec,
  agg_func = "sum")
}
\arguments{
\item{input_data}{The data containing period and DMA to which the footprint data needs to be added}

\item{metric_cols}{The columns for which the footprint data need to be added. Default value is media_names_vec}

\item{agg_func}{The function that needs to be used to generate the Footprint level aggregate data. Default is 'sum'}
}
\value{
Input data subset to the columns being aggregated,period and DMA with new records for the footprint level data
}
\description{
Function that adds aggregate data by Footprint by period
}
\details{
This function is used to add Footprint level aggregate data (Altice, Suddenlink and Optimum) as new DMAs
to the input_data, aggregated by period. \cr
Used in visualization functions to display metrics at the footprint level.
The default aggregation function used is sum. The aggregation function can be changed by passing the name of the aggregator as a string.
The function produces 3 new records by period with the data aggregated using group by period and aggregate function.
}
\note{
This function uses optimum_dma from the .GlobalEnv or the calling environment to compute the aggregate value for Optimum.
Does not remove the already existing Optimum data even if its a single Optimum DMA. If no metric_cols are provided, uses media_names_vec parameter from
.GlobalEnv or calling environment.
}
