% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{adStockGeom}
\alias{adStockGeom}
\title{Compute the adstock transformation on media data}
\usage{
adStockGeom(idx_media, alpha, input_data, max_lag = 14)
}
\arguments{
\item{idx_media}{ID of the media for which the adstock transformation needs to be computed}

\item{alpha}{Vector of the adstock parameter. Dimensions - n_media}

\item{max_lag}{Number of max lags applied for lag transformation of media data}

\item{input_media}{Normalized and lagged media data. Dimensions - (n_geo*nbr_periods)xn_mediaxn_lag}
}
\value{
Returns the vector of adstock transformed media data of length - nbr_periods*n_geo where nbr_periods is number of distinct periods of the data
}
\description{
Function to compute adstock transformation on media data
}
\details{
This function is used to apply the lagged transformation (adstock) effect of a single media across all DMA and period. \cr
Uses the estimated value of the alpha parameter and the lagged transformed input data to produce vector of weighted transformed output by media. \cr
The function computes the ratio of dot product of the weight and lagged media data and sum of the weights.
Weights being computed from the estimated alpha parameter of the media and number of lags
When using both hill and adstock transformation, this is first applied before applying the hill transformation.
}
