% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{Caps}
\alias{Caps}
\title{Convert Variable Names to Human Readable String}
\usage{
Caps(var_name)
}
\arguments{
\item{var_name}{Variable name with '_'}
}
\value{
Returns the Camel Cased human readable variable name after removing '_' from the name
}
\description{
Function to Convert Variable Names to Human Readable String
}
\details{
Use this function to convert any variable names to human readable strings by removing '_'. \cr
Removes the '_' in the variable name and converts the result to Camel Case. \cr
Best used for naming axis in plots. Replaces the '_' in the name with ' '. \cr
The function takes a single string as input. Use in conjunction with apply functions or mutate.
}
\examples{
Caps('sample_var_name')
Caps('camel_case')

}
