% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{clean_dma}
\alias{clean_dma}
\title{Clean DMA Names}
\usage{
clean_dma(dma_name)
}
\arguments{
\item{dma_name}{The name of the DMA that needs to be cleaned}
}
\value{
Returns the cleaned name of the input DMA
}
\description{
Function to clean the DMA names by removing special characters
}
\details{
Clean the names of the DMA by removing special characters and lowering the case. \cr
Use this function to clean and standardize DMA names before performing joins across different datasets. \cr
This function is applied to a single DMA name. So this needs to be used in conjunction with apply/mutate. \cr
}
