% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{exclude_zero_spend_weeks}
\alias{exclude_zero_spend_weeks}
\title{Exclude Zero spending periods from mean spending calculation}
\usage{
exclude_zero_spend_weeks(media_spend)
}
\arguments{
\item{media_spend}{The spend data for which mean needs to be computed excluding zero periods of spending}
}
\value{
mean_media_spend dataframe which is mean of spending values removing all periods of zero spending
}
\description{
Function to Calculate the mean media spending by DMA by excluding periods with zero spending
}
\details{
This function is used along with dplyr mutate to calculate the mean spending of a media by DMA if
zero spending periods are excluded. As some media have more zero spending periods than the others, the optimized results
might not be reflective of the actual spending levels. This function is used to calculate the mean spending values of medias by DMAs
which is used to compute the limits of simulation as well as optimization
}
\note{
This function is to be used as part of summarize_all and group by
}
