% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{format_money}
\alias{format_money}
\title{Add a Dollar Format to Money Fields}
\usage{
format_money(input_money, ndigits = 2)
}
\arguments{
\item{input_money}{Numerical value that needs to be dollar formatted}
}
\value{
Returns the dollar formatted value of input_money with ',' seperation for large values.
}
\description{
Function to change numerical values to dollar format
}
\details{
Use this function to format any money fields for pretty display. \cr
The function takes a single value as input, use this in conjunction with apply functions/mutate for multiple values. \cr
Any decimal values are rounded up to the nearest dollar value. \cr
The output is a character string with ',' seperation for larger values.
}
\examples{
format_money(3.5)
format_money(50000)

}
