% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{get_least_freq}
\alias{get_least_freq}
\title{Compute the Element with Least Frequency}
\usage{
get_least_freq(v)
}
\arguments{
\item{v}{Vector of categorical values}
}
\value{
Element with least frequency in the vector
}
\description{
Function to compute the element with the least frequency
}
\details{
This function is used to compute the element with least frequency of given categorical vector
}
\note{
When two elements have the same least frequency, it returns the first element.
}
\examples{
input_vec <- c(3,4,4,5)
get_least_freq(input_vec)

}
