% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_quarterly_mix_from_lists.R
\name{get_quarterly_mix}
\alias{get_quarterly_mix}
\title{Get Optimized Quarterly Mix}
\usage{
get_quarterly_mix(input_quarterly_budget_opt, input_quarterly_budget_sdl,
  sim_results_opt, sim_results_sdl, control_beta_mat, control_mat,
  control_names_vec, target_capita_norm_col, smoothened_res_OPT = F,
  smoothened_res_SDL = F, adjusted_res_OPT = F, adjusted_res_SDL = F,
  flight_media_vec_OPT = NULL, flight_media_vec_SDL = NULL,
  constr_media_vec_OPT = NULL, constr_media_vec_SDL = NULL,
  constr_media_budget_OPT = NULL, constr_media_budget_SDL = NULL,
  var_n_flight_period_OPT = NULL, var_n_flight_period_SDL = NULL,
  spend_media_vec = NULL, var_period = "Q1")
}
\arguments{
\item{input_quarterly_budget_opt}{Input quarterly Optimum budget}

\item{input_quarterly_budget_sdl}{Input quarterly Suddenlink budget}

\item{sim_results_opt}{Optimum results from the simulator}

\item{sim_results_sdl}{Suddenlink results from the simulator}

\item{control_beta_mat}{Beta estimates of the control variables}

\item{control_mat}{Actual,normalized values of the control variables from the preprocessing function}

\item{control_names_vec}{Names of the control variables used in the model. Has to be subset of the column names in control_mat}

\item{target_capita_norm_col}{Vector for denormalization of target value. 'subs' or 'non-subs'}

\item{constr_media_budget_OPT}{If there is a fixed Optimum budget of the constrained media to be added to the final optimal mix}

\item{constr_media_budget_SDL}{If there is a fixed Suddenlink budget of the constrained media to be added to the final optimal mix}

\item{var_period}{The quarter for which the optimized mix is being planned for. Uses the latest quarter data from the control data to calculate the target value of control}

\item{flight_media_vec}{Vector of the flight media. Currently supports single flight media. Default is NULL}

\item{const_media_vec}{Vector of constraint media. Use this if these media were excluded from the optimization across all DMA}

\item{var_n_flight_period}{Number of periods (weeks/months) of the flight media to be considered as 'On'}

\item{mean_spend_data}{Mean of spend data across all the weeks by DMA}
}
\value{
Formatted output list containing: \cr
Footprint metrics - Spend, target, CPMTGA, TGA, CPTGA \cr
DMA metrics - Spend, target, CPMTGA, TGA, CPTGA \cr

Unformatted raw output list containing: \cr
Quarterly budget mix by DMA and media \cr
Quarterly target values by DMA and media \cr
Target value across media by DMA \cr
}
\description{
Function to obtain the quarterly mix at given budget levels for given scenario
}
\details{
This function is used to combine the results obtained from the simulator for each footprint into one unified mix
at the quarterly level.

The simulator outputs the optimized budget mix by DMA at the period level (weekly/monthly) which need to be combined
into results at the quarterly level to be shared to the stakeholders. This function combines the results from the simulator
(output lists from the simulator) and generates the mix for a particular input Quarterly budget by Footprint.

This accounts for media that have been flighted and combines the on and off results from the simulator to get the final mix at the quarterly level.
}
