% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{get_quarters}
\alias{get_quarters}
\title{Add Quarter Columns to the Input Data}
\usage{
get_quarters(input_data)
}
\arguments{
\item{input_data}{Input data with the columns dma and period}
}
\value{
Returns the input data with the columns qy,year and quarter added.
}
\description{
Function to Add Quarter Columns to the Input Data
}
\details{
This function adds quarter related columns - 'quarter_year','year' and 'quarter' to any input data \cr
with the fields DMA and period. \cr
This function uses the global parameter nbr_period_q which is the number of periods in a quarter: \cr
If data is weekly, nbr_period_q = 13. If data is monthly, nbr_period_q = 3. \cr
This differs from the traditional quarter assigned by the datetime package as this function assumes each quarter has
fixed number of weeks (13). \cr
The year column is derived from the period. \cr
The period column is converted to date format using the as.Date function
}
\note{
nbr_period_q should be part of the GlobalEnv or any calling Env inorder to use this function.
}
\examples{
get_quarters(opt_data)
get_quarters(sdl_data)

}
