% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{get_target_dma_media}
\alias{get_target_dma_media}
\title{Compute the Actual Target for a Given Period}
\usage{
get_target_dma_media(input_spend, min_media_mat, max_media_mat,
  beta_media_mat, prospect_mat_media, media_name, dma_name, K, S,
  target_capita_norm_col, target_max_vec)
}
\arguments{
\item{input_spend}{The input spend for which the target value is being computed}

\item{min_media_mat}{The minimum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{max_media_mat}{The maximum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{beta_media_mat}{The matrix of beta coefficients of media by DMA estimated from the model. Dimensions - n_geo*n_media.}

\item{prospect_mat_media}{The matrix of subs/non_subs by DMA and media used to normalize media during preprocessing. Dimensions - n_geo*n_media}

\item{media_name}{The media for which the media target (without adstock) is being computed}

\item{dma_name}{The DMA for which the media target (without adstock) is being computed}

\item{K}{The vector of the Half-saturation parameter by media. Dimensions - n_media}

\item{S}{The vector of the Slope parameter by media. Dimensions - n_media}

\item{target_capita_norm_col}{The prospect(sub/non_sub) by DMA to de-normalize the target}

\item{target_max_vec}{The max of the target vector by DMA used for normalizing the target variable in preprocessing. Dimensions - n_geo}
}
\value{
target - The media target value for the given spend, DMA and media
}
\description{
Function that computes the model driven target value (no adstock) for given media and DMA
}
\details{
Compute the Media target for given spend
by DMA and media. Outputs the 
media target for the given spend by DMA and media.
}
