% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{getQuarterlydata}
\alias{getQuarterlydata}
\title{Aggregate weekly/monthly data to Quarterly level by DMA}
\usage{
getQuarterlydata(weekly_data)
}
\arguments{
\item{weekly_data}{The weekly/monthly data that needs to be rolled up to quarterly level by DMA.}
}
\value{
nbr_period_q*weekly_data aggregated at DMA level
}
\description{
Function to generate Quarterly level KPIs from weekly/monthly data by DMA
}
\details{
This function is used to generate Quarterly level KPIs by DMAs aggregated at the DMA level.
If the input data does not have a column name with 'dma', uses the dma_names_vec from the GlobalEnv to add the column for aggregation.
}
\note{
This function uses nbr_period_q from the .GlobalEnv or calling environment to generate the quarterly level metrics.
}
