% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{gross_add}
\alias{gross_add}
\title{Compute the Actual Target for a Given Period}
\usage{
gross_add(media, period)
}
\arguments{
\item{media}{Input media matrix/optimized activity from the optimizer}

\item{period}{Selected period for which the Control GA will be computed}
}
\value{
List of gross add objects like:
gross_add_control - Total Control GA for the selected period
gross_add_media - Total media GA
gross_add_total - Sum of media GA and control GA for the selected period
gross_add_total_sdl - Sum of media GA and control GA of Suddenlink DMAs
for the selected period
gross_add_total_opt - Sum of media GA and control GA of Optimum DMAs
for the selected period
gross_add_dma - Sum of media and control GA by DMA for the selected period
}
\description{
Function that computes the actual gross adds of a selected period by DMA and media
}
\details{
Compute the actual Gross adds - Media gross adds + Control gross adds for a given period
by DMA and media. Outputs a list of all the components that go into calculation of the
total gross adds of the period like control gross adds and media gross adds.
}
