% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{gross_add_control}
\alias{gross_add_control}
\title{Compute the Target Value for the Given Input Period by DMA}
\usage{
gross_add_control(Z, beta_control_df, control_names, data_XyZ,
  input_period, y_max, target_capita_norm_col, by_mean = F,
  nbr_period_q = 13)
}
\arguments{
\item{Z}{The preprocessed control matrix data}

\item{beta_control_df}{The posterior means of the control variables by DMA}

\item{control_names}{Names of the control variables}

\item{data_XyZ}{Actual data containing the period and DMA columns}

\item{input_period}{The period (quarter) for which the control values need to be computed}

\item{y_max}{The max value of the target by DMA to de-normalize the target}

\item{target_capita_norm_col}{The prospect(sub/non_sub) by DMA to de-normalize the data}

\item{by_mean}{Flag to compute the control value by mean or use the latest quarter (based on input). Default is F}

\item{nbr_period_q}{Number of periods that need to be considered in a quarter. Weekly - 13, Monthly - 3}
}
\value{
Returns the control target value by DMA and control variable
}
\description{
Function to compute the control portion of the target value by DMA and control variable.
}
\details{
Computes the control portion of the target for the input period.
Takes the control matrix and estimated values of the control parameters and computes the \cr
control target value. This will be added to the media portion of the target to obtain the total target for a given period (quarter). \cr
Use the by_mean parameter to compute the value based on mean or by pulling the last quarter of the input period. \cr
For example, if the input quarter is 'Q3' and the latest 'Q3' in the data is 'Q3 2019', it is used to calculate the control target value \cr
If by_mean is true, mean across all 'Q3' quarters is taken.
}
