% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{gross_add_weekly}
\alias{gross_add_weekly}
\title{Compute the Media Target of the Optimized Solution}
\usage{
gross_add_weekly(solution, min_media_mat, max_media_mat, ratio_dma, y_max,
  prospect_mat_media, beta_media_df, K, S, target_capita_norm_col)
}
\arguments{
\item{solution}{The output from the optimizer which is the optimized weekly activity}

\item{min_media_mat}{The minimum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{max_media_mat}{The maximum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{ratio_dma}{Spend to activity ratio dataframe of each media by DMA. Useful when modelling using activity media. Dimensions - n_dma*n_media}

\item{y_max}{The max of the target vector by DMA used for normalizing the target variable in preprocessing. Dimensions - n_geo}

\item{prospect_mat_media}{The matrix of subs/non_subs by DMA and media used to normalize media during preprocessing. Dimensions - n_geo*n_media}

\item{beta_media_df}{The matrix of beta coefficients of media by DMA estimated from the model. Dimensions - n_geo*n_media.}

\item{K}{The vector of the Half-saturation parameter by media. Dimensions - n_media}

\item{S}{The vector of the Slope parameter by media. Dimensions - n_media}

\item{target_capita_norm_col}{The prospect(sub/non_sub) by DMA to de-normalize the data}
}
\value{
Returns the media gross add matrix by media and DMA
}
\description{
Function to compute the media gross adds by DMA of the optimized solution
}
\details{
Computes the media gross adds of the optimized solution provided by the optimizer.
Takes the solution as the input and outputs the gross adds by media and by DMA using
the hill transformation. Used to show the weekly activity.
}
