% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{hill_second_derivative}
\alias{hill_second_derivative}
\title{Compute the second derivative of Hill function}
\usage{
hill_second_derivative(media, K, S)
}
\arguments{
\item{media}{Vector of normalized media data. Dimensions - n_media}

\item{K}{Vector of half saturation parameter. Dimensions - n_media}

\item{S}{Vector of slope parameter. Dimensions - n_media}
}
\value{
Returns the vector of second derivative of hill values for a DMA
}
\description{
Function to compute second derivative of the Hill function
}
\details{
This function is used to compute the second derivative (slope) of the hill function value of a single DMA across media. \cr
The function uses the estimated values of K(half saturation parameter) and S(Slope parameter). \cr
Takes a vector of media and vectors of estimated K and S as inputs and outputs the vector of second derivative. \cr
Used to evaluate the estimated values of K and S to visualize the S shaped curves for a DMA. \cr
The input media needs to be normalized before using this function.
}
\note{
This function cannot be used for a matrix of media values.
}
