% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{Mode}
\alias{Mode}
\title{Compute the Mode of Data}
\usage{
Mode(v)
}
\arguments{
\item{v}{Vector of categorical values}
}
\value{
Mode or element with highest frequency of the vector
}
\description{
Function to compute the mode of input data
}
\details{
This function is used to compute the mode of given categorical vector as R doesnt have an inbuilt mode function
}
\note{
When two elements have the same highest frequency, it returns the first element.
}
\examples{
input_vec <- c(3,4,4,5)
Mode(input_vec)

}
