% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_functions_mmm.R
\name{normalize_x}
\alias{normalize_x}
\title{Normalize DMA Media Data}
\usage{
normalize_x(x, min_media_mat, max_media_mat, prospect_mat_media)
}
\arguments{
\item{x}{Matrix of media data by DMA and media. Dimensions - n_geo*n_media}

\item{min_media_mat}{The minimum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{max_media_mat}{The maximum matrix of each media by DMA. Dimensions - n_geo*n_media}

\item{prospect_mat_media}{The matrix of subs/non_subs by DMA used to normalize media during preprocessing. Dimensions - n_geo*n_media}
}
\value{
Returns the normalized matrix of media data by DMA and media
}
\description{
Function to Normalize Media Data by Prospect and MinMax
}
\details{
This function is used to normalize all media data of all DMA by prospect (subs/non_subs) and
minmax by media. \cr
This function follows the normalization method specified under in the Google paper used to build the MMM model. \cr
The minmax can vary based on the normalization technique. It is obtained from the preprocessing module. \cr
The function outputs a matrix of normalized media data by DMA
}
