% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allocation_bar.r
\name{plot_allocation_bar}
\alias{plot_allocation_bar}
\title{Plot the quarterly Media Mix by DMA}
\usage{
plot_allocation_bar(DMA, quarterly_mix, readable_media_names = c("Radio",
  "TV", "Print", "DM", "Audio", "Display", "Video", "Branded Search",
  "Unbranded Search", "Social"))
}
\arguments{
\item{quarterly_mix}{Input spend dataframe for the media mix}

\item{readable_media_names}{Clean display names for graphs}

\item{dma}{Name of the dma for which the distribution needs to be plotted}
}
\value{
Returns the histogram of the media mix by DMA
}
\description{
Function to plot the distribution of media from the Quarterly Media Mix by DMA
}
\details{
Function that leverages the geom_bar of ggplot2 to show the distribution of the budget
among various media channels for the quarter.
Takes the input of the DMA and outputs the distribution of the media mix for the quarter
}
