% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_b2b_app.R
\name{run_b2b_app}
\alias{run_b2b_app}
\title{Get the Optimization Outputs for the Given Inputs}
\usage{
run_b2b_app(input)
}
\arguments{
\item{input}{JSON object containing all the inputs associated with the particular MMM}
}
\value{
For output - Footprint level metrics like target and spend
}
\description{
Function to get the distribution of media from the Quarterly Media Mix by footprint
}
\details{
Function that calls the get_quarterly_mix function and is used to expose the API for
obtaining the media mix.

Generates a summary table by footprint and formatted output by media

The input is input which is a generic JSON which is treated as a dataframe for generation
of the optimal mix.
}
\section{Note}{

For MMM Acquisition, for this function to work, please specify the following inputs:

Input_OPT_budget, \cr
Input_SDL_budget, \cr
Input_OPT_crosschannel_imp, \cr
Input_SDL_crosschannel_imp, \cr
Input_OPT_email_sent, \cr
Input_SDL_email_sent, \cr
input_period, \cr
}

